feat: Encrypted amount transfer split logic vs display
