+++ src/signCredentialDeployment.c
@@ -1,5 +1,5 @@
-#include <os.h>
-#include <os_io_seproxyhal.h>
+#include "os.h"
+#include "os_io_seproxyhal.h"
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
@@ -7,11 +7,12 @@
 #include "common/ui/display.h"
 #include "base58check.h"
 #include "cx.h"
-#include "menu.h"
-#include "responseCodes.h"
-#include "sign.h"
+#include "common/ui/menu.h"
+#include "common/responseCodes.h"
+#include "common/sign.h"
 #include "util.h"
 #include "signCredentialDeployment.h"
+#include "globals.h"
 
 static signCredentialDeploymentContext_t *ctx = &global.signCredentialDeploymentContext;
 static tx_state_t *tx_state = &global_tx_state;
